backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.14
langchain-groq==0.1.5
langchain-core==0.1.45
langchain-community==0.0.30
langsmith==0.1.17
chromadb==0.4.22
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
googlesearch-python==1.3.0
